TET-190: update `sync_company_with_dnb`
This method is being called both by the command through the job scheduler (`update_company_dnb_data`) and through an API post call (`DNBCompanyLinkView`). There was some logic to handle this in the method regarding retries, but as the job scheduler can handle retries automatically if the method errors, we no longer need this.