Management command to update duns_number
Part of work DEET is doing for Data Hub - implementing recommendations from
DnB, associating DnB numbers with Data Hub companies.

The immediate plan is to use this for a single company to verify the process.

This is likely to be followed up by using
https://github.com/uktrade/data-hub-api/blob/main/datahub/dbmaintenance/management/commands/update_company_dnb_data.py
to update data for the companies from DnB

This command won't be used for companies in DataHub that have the same DnB
number - i.e. duplicate companies. Dealing with that case is left to further
work.