fix: allow setting of null duns number
As part of dnb data cleanse using dnb's recommendations, we need
the ability to rollback changes which means setting duns number to
null rather than empty string. The empty string is not sufficient
because there can be at most 1 company with an empty string duns
number because of the unique constraint on that column.