Register CompanyReferral model with reversion
This is a stepping stone to make company merges undoable. Of the 6 models that
are changed by merging:

- Company
- CompanyReferral
- Contact
- Interaction
- InvestmentProject
- Order

CompanyReferral is the only one that was not tracked by reversion.

This is part of wider work by the DEET team to implement recommendations from
DnB in terms of data cleanup - and specifically about companies that are
duplicates and so would need merging.