Only add administrative areas for country and area matches
Some `AdministrativeArea`s do not have an abbreviated name associated
with them, so we must match by country and name. For example, Tamil Nadu
in India does not have an abbreviated name associated for us however
from DNB it has an abbreviated name of "TN" which would have been matched
to "Tennesse".