Create and apply migrations that have slipped through the net. (#5662)
These migrations change related field names which is a shortcut internal to Django ORM and does not actually modify the database, hence the ability for these to be missed. See the last paragraph of the [`The Commands` section](https://docs.djangoproject.com/en/5.1/topics/migrations/#the-commands) of the Django migration docs for more explanation as to why these migrations are created.