Re-add simplejson dependency and bump to 3.19.3
Upon removing the `simplejson` package and falling back to the built-in `json` library, it was found there were differences in the serialization of decimal numbers. This was causing the `test_gross_value_added_filters` test to fail when asserting the response of a project with max gross value added (99,999,989,999,999,999,990.000001).

Rather than spending time re-configuring the JSON encoder that Django uses, it was decided to keep the `simplejson` package and explicitly list it in the `requirements.txt` file.